requirements.txt
colour~=0.1
numpy~=2.1
pillow~=10.4
